// apps/web/src/api/types.ts
export type Media = {
  url: string;
  alternativeText?: string | null;
  width?: number | null;
  height?: number | null;
  mime?: string | null;
  formats?: Record<
    string,
    {
      url: string;
      width?: number | null;
      height?: number | null;
      mime?: string | null;
    }
  > | null;
};